Move EntryStore implementation into own module
