+++ p2panda-rs/src/test_utils/db/mod.rs
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod stores;
